Fix: 게시판 업데이트 현황 스켈레톤 UI 수정
