+++ components/common/UpdateBoard.jsx
@@ -1,24 +1,44 @@
 import {
-  Alert,
-  AlertDescription,
-  AlertIcon,
-  AlertTitle,
   Box,
   Heading,
+  Skeleton,
   Text,
   useBreakpointValue,
   useColorModeValue,
 } from '@chakra-ui/react';
-import React from 'react';
+import axios from 'axios';
+import { useEffect, useState } from 'react';
 
-import UpdateCard from '@/components/card/UpdateCard';
+import UpdateCardList from '@/components/card/UpdateCardList';
 import { darkMode, lightMode } from '@/styles/theme';
 
-const UpdateBoard = ({ last_update_info }) => {
+const UpdateBoard = () => {
   const color = useColorModeValue(lightMode.color, darkMode.color);
   const bg2 = useColorModeValue(lightMode.bg2, darkMode.bg2);
   const bg = useColorModeValue(lightMode.bg, darkMode.bg);
   const width = useBreakpointValue({ base: '100%', md: '100%' });
+
+  const [lastUpdateInfo, setLastUpdateInfo] = useState(null);
+  const [isLoading, setIsLoading] = useState(false);
+
+  useEffect(() => {
+    const fetchLastUpdateInfo = async () => {
+      try {
+        setIsLoading(true);
+        const response = await axios.get(
+          `${process.env.NEXT_PUBLIC_REDIRECT_URL}/last_update_info`
+        );
+        setLastUpdateInfo(response.data);
+      } catch (error) {
+        console.log('Error fetching last update info:', error);
+        // 오류 처리 로직
+      } finally {
+        setIsLoading(false);
+      }
+    };
+
+    fetchLastUpdateInfo();
+  }, []);
 
   return (
     <Box
@@ -54,29 +74,19 @@ const UpdateBoard = ({ last_update_info }) => {
           게시판 업데이트 현황
         </Heading>
       </Box>
-      <Box
-        display="flex"
-        flexDirection="column"
-        alignItems="center"
-        placeItems="center"
-        w="100%"
-        // p="1em"
-      >
-        {last_update_info?.map((update, index) => (
-          <UpdateCard key={index} update={update} />
-        ))}
-        {last_update_info === null ||
-          (last_update_info?.length === 0 && (
-            <Alert status="error" w="90%" borderRadius="1rem">
-              <AlertIcon />
-              <AlertTitle></AlertTitle>
-              <AlertDescription>
-                현재 서버와의 연결이 불안정합니다! 이용에 불편을 드려
-                죄송합니다. 빠른 시일 내에 해결하겠습니다.
-              </AlertDescription>
-            </Alert>
-          ))}
-      </Box>
+      {!isLoading ? (
+        <UpdateCardList updates={lastUpdateInfo} />
+      ) : (
+        <Box
+          maxW="700px"
+          mb="1rem"
+          w={width}
+          background="gray.700"
+          borderRadius="1rem"
+        >
+          <Skeleton height="600px" />
+        </Box>
+      )}
       <Box
         display="flex"
         flexDirection="column"
